Initial de0 nano port
This adds a far from complete port to the de0 nano board.
The de0 nano board has very few "built in" peripherals,
so extension boards for VGA, PS/2, sdcard, uart and joysticks
are needed.
How the external peripherals are connected are described in
fpga/de0_nano/extension_board.txt
Currently, I have only wired up the VGA, sdcard and uart,
so it's still a work in progress.
Also, the de0 nano board does not have any SRAM,
so the openrisc control processor shares the SDRAM
with minimig/tc68k.
The control processors memory space is located in
the upper 16MB of the 32MB SDRAM.
The old or1200 openrisc core has been swapped out for
the benefit of mor1kx, an other improved openrisc core.
Since we are now running out of SDRAM, mor1kx instruction
cache is enabled to avoid blocking tc68k too much.
Even so, it occupies less LUTs than or1200
(and there's still some things that can be disabled to make
it even smaller).

It is able to boot up a 1.2 kickstart ROM, but not much more
has been verified.